+++ pom.xml
@@ -9,7 +9,7 @@
         <version>1.8</version>
     </parent>
     <artifactId>sirius-search</artifactId>
-    <version>1.1</version>
+    <version>2.0-SNAPSHOT</version>
     <packaging>jar</packaging>
 
     <name>SIRIUS Search</name>
@@ -19,13 +19,18 @@
 
         <dependency>
             <groupId>com.scireum</groupId>
+            <artifactId>sirius-kernel</artifactId>
+            <version>2.2-SNAPSHOT</version>
+        </dependency>
+        <dependency>
+            <groupId>com.scireum</groupId>
             <artifactId>sirius-web</artifactId>
             <version>2.0.1</version>
         </dependency>
         <dependency>
             <groupId>com.scireum</groupId>
             <artifactId>sirius-kernel</artifactId>
-            <version>2.0</version>
+            <version>2.2-SNAPSHOT</version>
             <type>test-jar</type>
             <scope>test</scope>
         </dependency>
