+++ src/main/java/sirius/search/Index.java
@@ -8,175 +8,27 @@
 
 package sirius.search;
 
-import com.alibaba.fastjson.JSON;
-import com.alibaba.fastjson.JSONArray;
-import com.alibaba.fastjson.JSONObject;
-import com.google.common.base.Charsets;
-import com.google.common.collect.Lists;
-import com.google.common.collect.Maps;
-import com.google.common.io.CharStreams;
-import org.elasticsearch.action.admin.cluster.health.ClusterHealthResponse;
-import org.elasticsearch.action.admin.cluster.health.ClusterHealthStatus;
-import org.elasticsearch.action.admin.indices.create.CreateIndexResponse;
-import org.elasticsearch.action.admin.indices.delete.DeleteIndexResponse;
-import org.elasticsearch.action.admin.indices.exists.indices.IndicesExistsResponse;
-import org.elasticsearch.action.delete.DeleteRequestBuilder;
-import org.elasticsearch.action.get.GetResponse;
-import org.elasticsearch.action.index.IndexRequestBuilder;
-import org.elasticsearch.action.index.IndexResponse;
-import org.elasticsearch.client.Client;
-import org.elasticsearch.client.transport.TransportClient;
-import org.elasticsearch.common.settings.ImmutableSettings;
-import org.elasticsearch.common.settings.Settings;
-import org.elasticsearch.common.transport.InetSocketTransportAddress;
-import org.elasticsearch.index.engine.VersionConflictEngineException;
-import org.elasticsearch.node.Node;
-import org.elasticsearch.node.NodeBuilder;
-import sirius.kernel.Lifecycle;
-import sirius.kernel.Sirius;
-import sirius.kernel.async.Barrier;
-import sirius.kernel.async.CallContext;
-import sirius.kernel.async.ExecutionPoint;
 import sirius.kernel.async.Future;
-import sirius.kernel.async.Operation;
-import sirius.kernel.async.Tasks;
-import sirius.kernel.cache.Cache;
-import sirius.kernel.cache.CacheManager;
-import sirius.kernel.commons.Strings;
 import sirius.kernel.commons.Tuple;
-import sirius.kernel.commons.Wait;
-import sirius.kernel.commons.Watch;
-import sirius.kernel.di.std.ConfigValue;
 import sirius.kernel.di.std.Part;
-import sirius.kernel.di.std.Register;
-import sirius.kernel.health.Average;
-import sirius.kernel.health.Counter;
 import sirius.kernel.health.Exceptions;
 import sirius.kernel.health.HandledException;
-import sirius.kernel.health.Log;
-import sirius.kernel.health.metrics.MetricProvider;
-import sirius.kernel.health.metrics.MetricState;
-import sirius.kernel.health.metrics.MetricsCollector;
-import sirius.kernel.timer.EveryTenSeconds;
-import sirius.web.templates.Resource;
-import sirius.web.templates.Resources;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import java.io.File;
-import java.io.IOException;
-import java.io.InputStreamReader;
-import java.time.Duration;
-import java.util.Iterator;
-import java.util.List;
-import java.util.Map;
-import java.util.Timer;
-import java.util.TimerTask;
-import java.util.concurrent.TimeUnit;
 
 /**
  * Central access class to the persistence layer.
  * <p>
  * Provides CRUD access to the underlying ElasticSearch cluster.
+ *
+ * @deprecated use {@link IndexAccess}
  */
+@Deprecated
 public class Index {
 
-    /**
-     * Contains the default ID used by new entities
-     */
-    public static final String NEW = "new";
-
-    /**
-     * Contains the name of the ID field
-     */
-    public static final String ID_FIELD = "_id";
-
-    /**
-     * Async executor category for integrity check tasks
-     */
-    public static final String ASYNC_CATEGORY_INDEX_INTEGRITY = "index-ref-integrity";
-
-    /**
-     * Contains the database schema as expected by the java model
-     */
-    protected static Schema schema;
-
-    /**
-     * Logger used by this framework
-     */
-    public static final Log LOG = Log.get("index");
-
-    /**
-     * Contains the ElasticSearch client
-     */
-    private static Client client;
-    /**
-     * Contains the ES-Node if started using {@link #generateEmptyInMemoryInstance()}
-     */
-    private static Node inMemoryNode;
-
-    /**
-     * To support multiple installations in parallel, an indexPrefix can be supplied, which is added to each index
-     */
-    private static String indexPrefix;
-
-    /**
-     * Determines if the framework is already completely initialized
-     */
-    private static volatile boolean ready;
-
-    /**
-     * Contains a future which can be waited for
-     */
-    private static Future readyFuture = new Future();
-
-    /**
-     * Internal timer which is used to delay some actions. This is necessary, as ES takes up to one second to make a
-     * write visible to the next read
-     */
-    private static Timer delayLineTimer;
-
-    /**
-     * Can be used to cache frequently used entities.
-     */
-    private static Cache<String, Object> globalCache = CacheManager.createCache("entity-cache");
-
-    /**
-     * Used when optimistic lock tracing is enabled to record all changes
-     */
-    private static Map<String, IndexTrace> traces = Maps.newConcurrentMap();
-
-    /**
-     * Determines if optimistic lock errors should be traced
-     */
-    @ConfigValue("index.traceOptimisticLockErrors")
-    private static boolean traceOptimisticLockErrors;
-
-    /*
-     * Average query duration for statistical measures
-     */
-    private static Average queryDuration = new Average();
-    /*
-     * Counts how many threads used blockThreadForUpdate
-     */
-    private static Counter blocks = new Counter();
-    /*
-     * Counts how many delays where used
-     */
-    private static Counter delays = new Counter();
-    /*
-     * Counts how many optimistic lock errors occurred
-     */
-    private static Counter optimisticLockErrors = new Counter();
-
-    /**
-     * Can be used as routing value for one of the fetch methods to signal that no routing value is available
-     * and a lookup by select is deliberately called.
-     */
-    public static final String FETCH_DELIBERATELY_UNROUTED = "_DELIBERATELY_UNROUTED";
-
     @Part
-    private static Tasks tasks;
+    private static IndexAccess access;
 
     private Index() {
     }
@@ -201,32 +53,8 @@ public class Index {
                                                              @Nullable String id,
                                                              @Nonnull
                                                              com.google.common.cache.Cache<String, Object> cache) {
-        if (Strings.isEmpty(id)) {
-            return Tuple.create(null, false);
-        }
-        EntityDescriptor descriptor = Index.getDescriptor(type);
-
-        E value = (E) cache.getIfPresent(descriptor.getType() + "-" + id);
-        if (value != null) {
-            return Tuple.create(value, true);
-        }
-        if (descriptor.hasRouting()) {
-            if (Strings.isFilled(routing) && !FETCH_DELIBERATELY_UNROUTED.equals(routing)) {
-                value = find(routing, type, id);
-            } else {
-                if (FETCH_DELIBERATELY_UNROUTED.equals(routing)) {
-                    value = select(type).deliberatelyUnrouted().eq(ID_FIELD, id).queryFirst();
-                } else {
-                    value = select(type).eq(ID_FIELD, id).queryFirst();
-                }
-            }
-        } else {
-            value = find(type, id);
-        }
-        if (value != null) {
-            cache.put(descriptor.getType() + "-" + id, value);
-        }
-        return Tuple.create(value, false);
+        Exceptions.logDeprecatedMethodUse();
+        return access.fetch(routing, type, id, cache);
     }
 
     /**
@@ -246,7 +74,8 @@ public class Index {
                                                       @Nonnull Class<E> type,
                                                       @Nullable String id,
                                                       @Nonnull com.google.common.cache.Cache<String, Object> cache) {
-        return fetch(routing, type, id, cache).getFirst();
+        Exceptions.logDeprecatedMethodUse();
+        return access.fetchFromCache(routing, type, id, cache);
     }
 
     /**
@@ -266,30 +95,8 @@ public class Index {
     public static <E extends Entity> Tuple<E, Boolean> fetch(@Nullable String routing,
                                                              @Nonnull Class<E> type,
                                                              @Nullable String id) {
-        if (Strings.isEmpty(id)) {
-            return Tuple.create(null, false);
-        }
-        EntityDescriptor descriptor = Index.getDescriptor(type);
-
-        E value = (E) globalCache.get(descriptor.getType() + "-" + id);
-        if (value != null) {
-            return Tuple.create(value, true);
-        }
-        if (descriptor.hasRouting()) {
-            if (Strings.isFilled(routing) && !FETCH_DELIBERATELY_UNROUTED.equals(routing)) {
-                value = find(routing, type, id);
-            } else {
-                if (FETCH_DELIBERATELY_UNROUTED.equals(routing)) {
-                    value = select(type).deliberatelyUnrouted().eq(ID_FIELD, id).queryFirst();
-                } else {
-                    value = select(type).eq(ID_FIELD, id).queryFirst();
-                }
-            }
-        } else {
-            value = find(type, id);
-        }
-        globalCache.put(descriptor.getType() + "-" + id, value);
-        return Tuple.create(value, false);
+        Exceptions.logDeprecatedMethodUse();
+        return access.fetch(routing, type, id);
     }
 
     /**
@@ -307,7 +114,8 @@ public class Index {
     public static <E extends Entity> E fetchFromCache(@Nullable String routing,
                                                       @Nonnull Class<E> type,
                                                       @Nullable String id) {
-        return fetch(routing, type, id).getFirst();
+        Exceptions.logDeprecatedMethodUse();
+        return access.fetchFromCache(routing, type, id);
     }
 
     /**
@@ -316,45 +124,9 @@ public class Index {
      * @return the expected schema of the object model
      */
     public static Schema getSchema() {
-        return schema;
+        Exceptions.logDeprecatedMethodUse();
+        return access.getSchema();
     }
-
-    /**
-     * Used to delay an action for at least one second
-     */
-    private static class WaitingBlock {
-        private long waitline;
-        private Runnable cmd;
-        private CallContext context;
-
-        WaitingBlock(Runnable cmd) {
-            this.cmd = cmd;
-            this.waitline = System.currentTimeMillis() + 1000;
-            this.context = CallContext.getCurrent();
-        }
-
-        /**
-         * Determines if the action was delayed long enough
-         *
-         * @return <tt>true</tt> if the action was delayed long enough, <tt>false</tt> otherwise
-         */
-        public boolean isRunnable() {
-            return System.currentTimeMillis() > waitline;
-        }
-
-        /**
-         * Executes the action in its own executor
-         */
-        public void execute() {
-            CallContext.setCurrent(context);
-            tasks.executor("index-delay").fork(cmd);
-        }
-    }
-
-    /**
-     * Queue of actions which need to be delayed one second
-     */
-    private static final List<WaitingBlock> oneSecondDelayLine = Lists.newArrayList();
 
     /**
      * Adds an action to the delay line, which ensures that it is at least delayed for one second
@@ -362,93 +134,8 @@ public class Index {
      * @param cmd to command to be delayed
      */
     public static void callAfterUpdate(final Runnable cmd) {
-        synchronized (oneSecondDelayLine) {
-            if (oneSecondDelayLine.size() < 100) {
-                delays.inc();
-                oneSecondDelayLine.add(new WaitingBlock(cmd));
-                return;
-            }
-        }
-        blockThreadForUpdate();
-        cmd.run();
-    }
-
-    /**
-     * Provides usage metrics for elasticsearch and the query system.
-     */
-    @Register
-    public static class IndexReport implements MetricProvider {
-
-        @Override
-        public void gather(MetricsCollector collector) {
-            if (!isReady()) {
-                return;
-            }
-            ClusterHealthResponse res = getClient().admin().cluster().prepareHealth().execute().actionGet();
-            collector.metric("ES-Nodes", res.getNumberOfNodes(), null, asMetricState(res.getStatus()));
-            collector.metric("ES-InitializingShards",
-                             res.getInitializingShards(),
-                             null,
-                             res.getInitializingShards() > 0 ? MetricState.YELLOW : MetricState.GRAY);
-            collector.metric("ES-RelocatingShards",
-                             res.getRelocatingShards(),
-                             null,
-                             res.getRelocatingShards() > 0 ? MetricState.YELLOW : MetricState.GRAY);
-            collector.metric("ES-UnassignedShards",
-                             res.getUnassignedShards(),
-                             null,
-                             res.getUnassignedShards() > 0 ? MetricState.RED : MetricState.GRAY);
-            collector.metric("index-delay-line", "ES-DelayLine", oneSecondDelayLine.size(), null);
-            collector.differentialMetric("index-blocks", "index-blocks", "ES-DelayBlocks", blocks.getCount(), "/min");
-            collector.differentialMetric("index-delays", "index-delays", "ES-Delays", delays.getCount(), "/min");
-            collector.differentialMetric("index-locking-errors",
-                                         "index-locking-errors",
-                                         "ES-OptimisticLock-Errors",
-                                         optimisticLockErrors.getCount(),
-                                         "/min");
-            collector.metric("index-queryDuration", "ES-QueryDuration", queryDuration.getAndClearAverage(), "ms");
-            collector.differentialMetric("index-queries",
-                                         "index-queries",
-                                         "ES-Queries",
-                                         queryDuration.getCount(),
-                                         "/min");
-        }
-
-        private MetricState asMetricState(ClusterHealthStatus status) {
-            if (status == ClusterHealthStatus.GREEN) {
-                return MetricState.GRAY;
-            } else if (status == ClusterHealthStatus.YELLOW) {
-                return MetricState.YELLOW;
-            } else {
-                return MetricState.RED;
-            }
-        }
-    }
-
-    /**
-     * Implementation of the timer which handles delayed actions.
-     */
-    protected static class DelayLineHandler extends TimerTask {
-
-        @Override
-        public void run() {
-            try {
-                synchronized (oneSecondDelayLine) {
-                    Iterator<WaitingBlock> iter = oneSecondDelayLine.iterator();
-                    while (iter.hasNext()) {
-                        WaitingBlock next = iter.next();
-                        if (next.isRunnable()) {
-                            next.execute();
-                            iter.remove();
-                        } else {
-                            return;
-                        }
-                    }
-                }
-            } catch (Throwable e) {
-                Exceptions.handle(LOG, e);
-            }
-        }
+        Exceptions.logDeprecatedMethodUse();
+        access.callAfterUpdate(cmd);
     }
 
     /**
@@ -458,8 +145,8 @@ public class Index {
      * is absolutely necessary.
      */
     public static void blockThreadForUpdate() {
-        blocks.inc();
-        Wait.seconds(1);
+        Exceptions.logDeprecatedMethodUse();
+        access.blockThreadForUpdate();
     }
 
     /**
@@ -468,7 +155,8 @@ public class Index {
      * @return <tt>true</tt> if the framework is completely initialized, <tt>false</tt> otherwise
      */
     public static boolean isReady() {
-        return ready;
+        Exceptions.logDeprecatedMethodUse();
+        return access.isReady();
     }
 
     /**
@@ -477,148 +165,16 @@ public class Index {
      * @return a future which is fullfilled once the index is ready
      */
     public static Future ready() {
-        return readyFuture;
+        Exceptions.logDeprecatedMethodUse();
+        return access.ready();
     }
 
     /**
      * Blocks the calling thread until the index is ready.
      */
     public static void waitForReady() {
-        if (!Index.isReady()) {
-            try {
-                Barrier b = Barrier.create();
-                b.add(readyFuture);
-                b.await();
-            } catch (InterruptedException e) {
-                Exceptions.handle(e);
-            }
-        }
-    }
-
-    /**
-     * Removes outdated traces used to discover optimistic lock errors
-     */
-    @Register
-    public static class OptimisticLockTracer implements EveryTenSeconds {
-
-        @Part
-        private Tasks tasks;
-
-        @Override
-        public void runTimer() throws Exception {
-            if (traceOptimisticLockErrors) {
-                tasks.defaultExecutor().fork(this::cleanOldRecordings);
-            }
-        }
-
-        private void cleanOldRecordings() {
-            long limit = System.currentTimeMillis() - 10_000;
-            Iterator<IndexTrace> iter = traces.values().iterator();
-            while (iter.hasNext()) {
-                if (iter.next().timestamp < limit) {
-                    iter.remove();
-                }
-            }
-        }
-    }
-
-    /**
-     * Starts and stops the elasticsearch client.
-     */
-    @Register(classes = Lifecycle.class)
-    public static class IndexLifecycle implements Lifecycle {
-
-        @Override
-        public int getPriority() {
-            return 75;
-        }
-
-        @Override
-        public void started() {
-            if (Strings.isEmpty(Sirius.getConfig().getString("index.type"))) {
-                LOG.INFO("ElasticSearch is disabled! (index.type is not set)");
-                return;
-            }
-
-            Operation.cover("index", () -> "IndexLifecycle.startClient", Duration.ofSeconds(15), this::startClient);
-
-            // Setup index
-            indexPrefix = Sirius.getConfig().getString("index.prefix");
-            if (!indexPrefix.endsWith("-")) {
-                indexPrefix = indexPrefix + "-";
-            }
-
-            schema = new Schema();
-            schema.load();
-
-            Operation.cover("index",
-                            () -> "IndexLifecycle.updateMappings",
-                            Duration.ofSeconds(30),
-                            this::updateMappings);
-
-            ready = true;
-            readyFuture.success();
-
-            delayLineTimer = new Timer("index-delay");
-            delayLineTimer.schedule(new DelayLineHandler(), 1000, 1000);
-        }
-
-        private void updateMappings() {
-            boolean updateSchema = Sirius.getConfig().getBoolean("index.updateSchema")
-                                   || "in-memory".equalsIgnoreCase(Sirius.getConfig().getString("index.type"));
-
-            if (updateSchema) {
-                for (String msg : schema.createMappings()) {
-                    LOG.INFO(msg);
-                }
-            }
-        }
-
-        private void startClient() {
-            if ("embedded".equalsIgnoreCase(Sirius.getConfig().getString("index.type"))) {
-                LOG.INFO("Starting Embedded Elasticsearch...");
-                client = NodeBuilder.nodeBuilder().data(true).local(true).build().client();
-            } else if ("in-memory".equalsIgnoreCase(Sirius.getConfig().getString("index.type"))) {
-                LOG.INFO("Starting In-Memory Elasticsearch...");
-                generateEmptyInMemoryInstance();
-            } else {
-                LOG.INFO("Connecting to Elasticsearch cluster '%s' via '%s'...",
-                         Sirius.getConfig().getString("index.cluster"),
-                         Sirius.getConfig().getString("index.host"));
-                Settings settings = ImmutableSettings.settingsBuilder()
-                                                     .put("cluster.name", Sirius.getConfig().getString("index.cluster"))
-                                                     .build();
-                TransportClient transportClient = new TransportClient(settings);
-                transportClient.addTransportAddress(new InetSocketTransportAddress(Sirius.getConfig()
-                                                                                         .getString("index.host"),
-                                                                                   Sirius.getConfig()
-                                                                                         .getInt("index.port")));
-                client = transportClient;
-            }
-        }
-
-        @Override
-        public void stopped() {
-            if (delayLineTimer != null) {
-                delayLineTimer.cancel();
-            }
-        }
-
-        @Override
-        public void awaitTermination() {
-            // We wait until this last call before we cut the connection to the database (elasticsearch) to permit
-            // other stopping lifecycles access until the very end...
-            ready = false;
-            client.close();
-            if (inMemoryNode != null) {
-                inMemoryNode.close();
-            }
-        }
-
-        @Override
-        public String getName() {
-            return "index (ElasticSearch)";
-        }
+        Exceptions.logDeprecatedMethodUse();
+        access.waitForReady();
     }
 
     /**
@@ -629,40 +185,8 @@ public class Index {
      *                          lock error.
      */
     public static void retry(UnitOfWork uow) {
-        int retries = 3;
-        while (retries > 0) {
-            retries--;
-            try {
-                uow.execute();
-                return;
-            } catch (OptimisticLockException e) {
-                LOG.FINE(e);
-                if (Sirius.isDev()) {
-                    LOG.INFO("Retrying due to optimistic lock: %s", e);
-                }
-                if (retries <= 0) {
-                    throw Exceptions.handle()
-                                    .withSystemErrorMessage(
-                                            "Failed to update an entity after re-trying a unit of work several times: %s (%s)")
-                                    .error(e)
-                                    .to(LOG)
-                                    .handle();
-                }
-                // Wait 0, 500ms, 1000ms
-                Wait.millis((2 - retries) * 500);
-                // Wait 0..500ms in 50% of all calls...
-                Wait.randomMillis(-500, 500);
-            } catch (HandledException e) {
-                throw e;
-            } catch (Throwable e) {
-                throw Exceptions.handle()
-                                .withSystemErrorMessage(
-                                        "An unexpected exception occurred while executing a unit of work: %s (%s)")
-                                .error(e)
-                                .to(LOG)
-                                .handle();
-            }
-        }
+        Exceptions.logDeprecatedMethodUse();
+        access.retry(uow);
     }
 
     /**
@@ -672,18 +196,8 @@ public class Index {
      * @return <tt>true</tt> if the given index exists, <tt>false</tt> otherwise
      */
     public static boolean existsIndex(String name) {
-        String index = getIndexName(name);
-        try {
-            IndicesExistsResponse res =
-                    Index.getClient().admin().indices().prepareExists(index).execute().get(10, TimeUnit.SECONDS);
-            return res.isExists();
-        } catch (Throwable e) {
-            throw Exceptions.handle()
-                            .to(LOG)
-                            .error(e)
-                            .withSystemErrorMessage("Failed to check existence of index: %s - %s (%s)", index)
-                            .handle();
-        }
+        Exceptions.logDeprecatedMethodUse();
+        return access.existsIndex(name);
     }
 
     /**
@@ -692,36 +206,8 @@ public class Index {
      * @param name the name of the index. The index prefix of the current system will be added automatically
      */
     public static void ensureIndexExists(String name) {
-        try {
-            IndicesExistsResponse res = Index.getClient()
-                                             .admin()
-                                             .indices()
-                                             .prepareExists(getIndexName(name))
-                                             .execute()
-                                             .get(10, TimeUnit.SECONDS);
-            if (!res.isExists()) {
-                CreateIndexResponse createResponse = Index.getClient()
-                                                          .admin()
-                                                          .indices()
-                                                          .prepareCreate(getIndexName(name))
-                                                          .execute()
-                                                          .get(10, TimeUnit.SECONDS);
-                if (!createResponse.isAcknowledged()) {
-                    throw Exceptions.handle()
-                                    .to(LOG)
-                                    .withSystemErrorMessage("Cannot create index: %s", getIndexName(name))
-                                    .handle();
-                } else {
-                    Index.blockThreadForUpdate();
-                }
-            }
-        } catch (Throwable e) {
-            throw Exceptions.handle()
-                            .to(LOG)
-                            .error(e)
-                            .withSystemErrorMessage("Cannot create index: %s - %s (%s)", getIndexName(name))
-                            .handle();
-        }
+        Exceptions.logDeprecatedMethodUse();
+        access.ensureIndexExists(name);
     }
 
     /**
@@ -730,26 +216,8 @@ public class Index {
      * @param name the name of the index. The index prefix of the current system will be added automatically
      */
     public static void deleteIndex(String name) {
-        try {
-            DeleteIndexResponse res = Index.getClient()
-                                           .admin()
-                                           .indices()
-                                           .prepareDelete(getIndexName(name))
-                                           .execute()
-                                           .get(10, TimeUnit.SECONDS);
-            if (!res.isAcknowledged()) {
-                throw Exceptions.handle()
-                                .to(LOG)
-                                .withSystemErrorMessage("Cannot delete index: %s", getIndexName(name))
-                                .handle();
-            }
-        } catch (Throwable e) {
-            throw Exceptions.handle()
-                            .to(LOG)
-                            .error(e)
-                            .withSystemErrorMessage("Cannot delete index: %s - %s (%s)", getIndexName(name))
-                            .handle();
-        }
+        Exceptions.logDeprecatedMethodUse();
+        access.deleteIndex(name);
     }
 
     /**
@@ -761,28 +229,8 @@ public class Index {
      * @param <E>   the generic of <tt>type</tt>
      */
     public static <E extends Entity> void addMapping(String index, Class<E> type) {
-        try {
-            EntityDescriptor desc = schema.getDescriptor(type);
-            Index.getClient()
-                 .admin()
-                 .indices()
-                 .preparePutMapping(index)
-                 .setType(desc.getType())
-                 .setSource(desc.createMapping())
-                 .execute()
-                 .get(10, TimeUnit.SECONDS);
-        } catch (Throwable ex) {
-            while (ex.getCause() != null && ex.getCause() != ex) {
-                ex = ex.getCause();
-            }
-            throw Exceptions.handle()
-                            .to(LOG)
-                            .error(ex)
-                            .withSystemErrorMessage("Cannot create mapping %s in index: %s - %s (%s)",
-                                                    type.getSimpleName(),
-                                                    index)
-                            .handle();
-        }
+        Exceptions.logDeprecatedMethodUse();
+        access.addMapping(index, type);
     }
 
     /**
@@ -793,12 +241,8 @@ public class Index {
      * @return the stored entity (with a filled ID etc.)
      */
     public static <E extends Entity> E create(E entity) {
-        try {
-            return update(entity, false, true);
-        } catch (OptimisticLockException e) {
-            // Should never happen - but who knows...
-            throw Exceptions.handle(LOG, e);
-        }
+        Exceptions.logDeprecatedMethodUse();
+        return access.create(entity);
     }
 
     /**
@@ -814,7 +258,8 @@ public class Index {
      *                                 by the entity to be saved
      */
     public static <E extends Entity> E tryUpdate(E entity) throws OptimisticLockException {
-        return update(entity, true, false);
+        Exceptions.logDeprecatedMethodUse();
+        return access.tryUpdate(entity);
     }
 
     /**
@@ -828,18 +273,8 @@ public class Index {
      * @return the updated entity
      */
     public static <E extends Entity> E update(E entity) {
-        try {
-            return update(entity, true, false);
-        } catch (OptimisticLockException e) {
-            reportClash(entity);
-            throw Exceptions.handle()
-                            .to(LOG)
-                            .error(e)
-                            .withSystemErrorMessage("Failed to update '%s' (%s): %s (%s)",
-                                                    entity.toDebugString(),
-                                                    entity.getId())
-                            .handle();
-        }
+        Exceptions.logDeprecatedMethodUse();
+        return access.update(entity);
     }
 
     /**
@@ -853,127 +288,8 @@ public class Index {
      * @return the updated entity
      */
     public static <E extends Entity> E override(E entity) {
-        try {
-            return update(entity, false, false);
-        } catch (OptimisticLockException e) {
-            // Should never happen as version checking is disabled....
-            throw Exceptions.handle(LOG, e);
-        }
-    }
-
-    /**
-     * Internal save method used by {@link #create(Entity)}, {@link #tryUpdate(Entity)}, {@link #update(Entity)}
-     * and {@link #override(Entity)}
-     *
-     * @param entity              the entity to save
-     * @param performVersionCheck determines if change tracking will be enabled
-     * @param forceCreate         determines if a new entity should be created
-     * @param <E>                 the type of the entity to update
-     * @return the saved entity
-     * @throws OptimisticLockException if change tracking is enabled and an intermediary change took place
-     */
-    protected static <E extends Entity> E update(final E entity,
-                                                 final boolean performVersionCheck,
-                                                 final boolean forceCreate) throws OptimisticLockException {
-        try {
-            final Map<String, Object> source = Maps.newTreeMap();
-            entity.beforeSave();
-            EntityDescriptor descriptor = getDescriptor(entity.getClass());
-            descriptor.writeTo(entity, source);
-
-            if (LOG.isFINE()) {
-                LOG.FINE("SAVE[CREATE: %b, LOCK: %b]: %s.%s: %s",
-                         forceCreate,
-                         performVersionCheck,
-                         getIndex(entity),
-                         descriptor.getType(),
-                         Strings.join(source));
-            }
-
-            String id = entity.getId();
-            if (NEW.equals(id)) {
-                id = null;
-            }
-            if (Strings.isEmpty(id)) {
-                id = entity.computePossibleId();
-            }
-
-            IndexRequestBuilder irb = getClient().prepareIndex(getIndex(entity), descriptor.getType(), id)
-                                                 .setCreate(forceCreate)
-                                                 .setSource(source);
-            if (!entity.isNew() && performVersionCheck) {
-                irb.setVersion(entity.getVersion());
-            }
-            if (descriptor.hasRouting()) {
-                Object routingKey = descriptor.getProperty(descriptor.getRouting()).writeToSource(entity);
-                if (Strings.isEmpty(routingKey)) {
-                    LOG.WARN("Updating an entity of type %s (%s) without routing information! Location: %s",
-                             entity.getClass().getName(),
-                             entity.getId(),
-                             ExecutionPoint.snapshot());
-                } else {
-                    irb.setRouting(String.valueOf(routingKey));
-                }
-            }
-
-            return executeUpdate(entity, descriptor, irb);
-        } catch (VersionConflictEngineException e) {
-            if (LOG.isFINE()) {
-                LOG.FINE("Version conflict on updating: %s", entity);
-            }
-            optimisticLockErrors.inc();
-            throw new OptimisticLockException(e, entity);
-        } catch (Throwable e) {
-            throw Exceptions.handle()
-                            .to(LOG)
-                            .error(e)
-                            .withSystemErrorMessage("Failed to update '%s' (%s): %s (%s)",
-                                                    entity.toDebugString(),
-                                                    entity.getId())
-                            .handle();
-        }
-    }
-
-    private static <E extends Entity> E executeUpdate(E entity, EntityDescriptor descriptor, IndexRequestBuilder irb) {
-        Watch w = Watch.start();
-        IndexResponse indexResponse = irb.execute().actionGet();
-        if (LOG.isFINE()) {
-            LOG.FINE("SAVE: %s.%s: %s (%d) SUCCEEDED",
-                     getIndex(entity),
-                     descriptor.getType(),
-                     indexResponse.getId(),
-                     indexResponse.getVersion());
-        }
-        entity.id = indexResponse.getId();
-        entity.version = indexResponse.getVersion();
-        entity.afterSave();
-        queryDuration.addValue(w.elapsedMillis());
-        w.submitMicroTiming("ES", "UPDATE " + entity.getClass().getName());
-        traceChange(entity);
-        return entity;
-    }
-
-    /**
-     * Returns the descriptor for the given class.
-     *
-     * @param clazz the class which descriptor is request
-     * @return the descriptor for the given class
-     */
-    public static EntityDescriptor getDescriptor(Class<? extends Entity> clazz) {
-        if (!ready) {
-            throw Exceptions.handle().to(LOG).withSystemErrorMessage("Index is not ready yet.").handle();
-        }
-        return schema.getDescriptor(clazz);
-    }
-
-    /**
-     * Returns the class for the given type name.
-     *
-     * @param name the name of the type which class is requested
-     * @return the class which is associated with the given type
-     */
-    public static Class<? extends Entity> getType(String name) {
-        return schema.getType(name);
+        Exceptions.logDeprecatedMethodUse();
+        return access.override(entity);
     }
 
     /**
@@ -985,7 +301,8 @@ public class Index {
      * @return the entity of the given class with the given id or <tt>null</tt> if no such entity exists
      */
     public static <E extends Entity> E find(final Class<E> clazz, String id) {
-        return find(null, null, clazz, id);
+        Exceptions.logDeprecatedMethodUse();
+        return access.find(null, null, clazz, id);
     }
 
     /**
@@ -998,7 +315,8 @@ public class Index {
      * @return the entity of the given class with the given id or <tt>null</tt> if no such entity exists
      */
     public static <E extends Entity> E find(String routing, final Class<E> clazz, String id) {
-        return find(null, routing, clazz, id);
+        Exceptions.logDeprecatedMethodUse();
+        return access.find(null, routing, clazz, id);
     }
 
     /**
@@ -1016,82 +334,8 @@ public class Index {
                                             @Nullable String routing,
                                             @Nonnull final Class<E> clazz,
                                             String id) {
-        try {
-            if (Strings.isEmpty(id)) {
-                return null;
-            }
-            if (NEW.equals(id)) {
-                E e = clazz.newInstance();
-                e.setId(NEW);
-                return e;
-            }
-            if (index == null) {
-                index = getIndex(clazz);
-            } else {
-                index = getIndexName(index);
-            }
-            EntityDescriptor descriptor = getDescriptor(clazz);
-            if (LOG.isFINE()) {
-                LOG.FINE("FIND: %s.%s: %s", index, descriptor.getType(), id);
-            }
-            Watch w = Watch.start();
-            try {
-                verifyRoutingForFind(routing, clazz, id, descriptor);
-                GetResponse res = getClient().prepareGet(index, descriptor.getType(), id)
-                                             .setPreference("_primary")
-                                             .setRouting(routing)
-                                             .execute()
-                                             .actionGet();
-                if (!res.isExists()) {
-                    if (LOG.isFINE()) {
-                        LOG.FINE("FIND: %s.%s: NOT FOUND", index, descriptor.getType());
-                    }
-                    return null;
-                } else {
-                    E entity = clazz.newInstance();
-                    entity.initSourceTracing();
-                    entity.setId(res.getId());
-                    entity.setVersion(res.getVersion());
-                    descriptor.readSource(entity, res.getSource());
-                    if (LOG.isFINE()) {
-                        LOG.FINE("FIND: %s.%s: FOUND: %s", index, descriptor.getType(), Strings.join(res.getSource()));
-                    }
-                    return entity;
-                }
-            } finally {
-                queryDuration.addValue(w.elapsedMillis());
-                w.submitMicroTiming("ES", "UPDATE " + clazz.getName());
-            }
-        } catch (Throwable t) {
-            throw Exceptions.handle()
-                            .to(LOG)
-                            .error(t)
-                            .withSystemErrorMessage("Failed to find '%s' (%s): %s (%s)", id, clazz.getName())
-                            .handle();
-        }
-    }
-
-    private static <E extends Entity> void verifyRoutingForFind(@Nullable String routing,
-                                                                @Nonnull Class<E> clazz,
-                                                                String id,
-                                                                EntityDescriptor descriptor) {
-        if (descriptor.hasRouting() && routing == null) {
-            Exceptions.handle()
-                      .to(LOG)
-                      .withSystemErrorMessage(
-                              "Trying to FIND an entity of type %s (with id %s) without providing a routing! "
-                              + "This will most probably FAIL!",
-                              clazz.getName(),
-                              id)
-                      .handle();
-        } else if (!descriptor.hasRouting() && routing != null) {
-            Exceptions.handle()
-                      .to(LOG)
-                      .withSystemErrorMessage("Trying to FIND an entity of type %s (with id %s) with a routing "
-                                              + "- but entity has no routing attribute (in @Indexed)! "
-                                              + "This will most probably FAIL!", clazz.getName(), id)
-                      .handle();
-        }
+        Exceptions.logDeprecatedMethodUse();
+        return access.find(index, routing, clazz, id);
     }
 
     /**
@@ -1109,20 +353,8 @@ public class Index {
     @SuppressWarnings("unchecked")
     @Nullable
     public static <T extends Entity> T refreshOrNull(@Nullable T entity) {
-        if (entity == null) {
-            return null;
-        }
-        if (entity.isNew()) {
-            return entity;
-        }
-        Class<T> type = (Class<T>) entity.getClass();
-        EntityDescriptor descriptor = getDescriptor(type);
-        if (descriptor.hasRouting()) {
-            Object routing = descriptor.getProperty(descriptor.getRouting()).writeToSource(entity);
-            return Index.find(routing == null ? null : routing.toString(), type, entity.getId());
-        } else {
-            return Index.find(type, entity.getId());
-        }
+        Exceptions.logDeprecatedMethodUse();
+        return access.refreshOrNull(entity);
     }
 
     /**
@@ -1138,18 +370,8 @@ public class Index {
      */
     @Nullable
     public static <T extends Entity> T refreshOrFail(@Nullable T entity) {
-        T freshEntity = refreshOrNull(entity);
-        if (entity != null && freshEntity == null) {
-            throw Exceptions.handle()
-                            .to(LOG)
-                            .withSystemErrorMessage("Failed to refresh the entity '%s' of type %s with id '%s'",
-                                                    entity,
-                                                    entity.getClass().getSimpleName(),
-                                                    entity.getId())
-                            .handle();
-        } else {
-            return freshEntity;
-        }
+        Exceptions.logDeprecatedMethodUse();
+        return access.refreshOrFail(entity);
     }
 
     /**
@@ -1164,50 +386,8 @@ public class Index {
      */
     @Nullable
     public static <T extends Entity> T refreshIfPossible(@Nullable T entity) {
-        T freshEntity = refreshOrNull(entity);
-        if (freshEntity == null) {
-            return entity;
-        } else {
-            return freshEntity;
-        }
-    }
-
-    /**
-     * Returns the name of the index associated with the given entity.
-     *
-     * @param entity the entity which index is requested
-     * @param <E>    the type of the entity to get the index for
-     * @return the index name associated with the given entity
-     */
-    protected static <E extends Entity> String getIndex(E entity) {
-        if (entity != null) {
-            String result = entity.getIndex();
-            if (result != null) {
-                return getIndexName(result);
-            }
-        }
-        return getIndexName(getDescriptor(entity.getClass()).getIndex());
-    }
-
-    /**
-     * Returns the name of the index associated with the given class.
-     *
-     * @param clazz the entity type which index is requested
-     * @param <E>   the type of the entity to get the index for
-     * @return the index name associated with the given class
-     */
-    protected static <E extends Entity> String getIndex(Class<E> clazz) {
-        return getIndexName(getDescriptor(clazz).getIndex());
-    }
-
-    /**
-     * Qualifies the given index name with the prefix.
-     *
-     * @param name the index name to qualify
-     * @return the qualified name of the given index name
-     */
-    protected static String getIndexName(String name) {
-        return indexPrefix + name;
+        Exceptions.logDeprecatedMethodUse();
+        return access.refreshIfPossible(entity);
     }
 
     /**
@@ -1218,7 +398,8 @@ public class Index {
      * @throws OptimisticLockException if the entity was modified since the last read
      */
     public static <E extends Entity> void tryDelete(E entity) throws OptimisticLockException {
-        delete(entity, false);
+        Exceptions.logDeprecatedMethodUse();
+        access.tryDelete(entity);
     }
 
     /**
@@ -1230,17 +411,8 @@ public class Index {
      * @param <E>    the type of the entity to delete
      */
     public static <E extends Entity> void delete(E entity) {
-        try {
-            delete(entity, false);
-        } catch (OptimisticLockException e) {
-            throw Exceptions.handle()
-                            .to(LOG)
-                            .error(e)
-                            .withSystemErrorMessage("Failed to delete '%s' (%s): %s (%s)",
-                                                    entity.toDebugString(),
-                                                    entity.getId())
-                            .handle();
-        }
+        Exceptions.logDeprecatedMethodUse();
+        access.delete(entity);
     }
 
     /**
@@ -1251,143 +423,8 @@ public class Index {
      * @param <E>    the type of the entity to delete
      */
     public static <E extends Entity> void forceDelete(E entity) {
-        try {
-            delete(entity, true);
-        } catch (OptimisticLockException e) {
-            // Should never happen as version checking is disabled....
-            throw Exceptions.handle(LOG, e);
-        }
-    }
-
-    /**
-     * Handles all kinds of deletes
-     *
-     * @param entity the entity to delete
-     * @param force  determines whether optimistic locking is suppressed (<tt>true</tt>) or not
-     * @param <E>    the type of the entity to delete
-     * @throws OptimisticLockException if the entity was changed since the last read
-     */
-    protected static <E extends Entity> void delete(final E entity, final boolean force)
-            throws OptimisticLockException {
-        try {
-            if (entity.isNew()) {
-                return;
-            }
-            EntityDescriptor descriptor = getDescriptor(entity.getClass());
-            if (LOG.isFINE()) {
-                LOG.FINE("DELETE[FORCE: %b]: %s.%s: %s",
-                         force,
-                         getIndex(entity.getClass()),
-                         descriptor.getType(),
-                         entity.getId());
-            }
-            entity.beforeDelete();
-            Watch w = Watch.start();
-            DeleteRequestBuilder drb =
-                    getClient().prepareDelete(getIndex(entity), descriptor.getType(), entity.getId());
-            if (!force) {
-                drb.setVersion(entity.getVersion());
-            }
-            if (descriptor.hasRouting()) {
-                Object routingKey = descriptor.getProperty(descriptor.getRouting()).writeToSource(entity);
-                if (Strings.isEmpty(routingKey)) {
-                    LOG.WARN("Deleting an entity of type %s (%s) without routing information! Location: %s",
-                             entity.getClass().getName(),
-                             entity.getId(),
-                             ExecutionPoint.snapshot());
-                } else {
-                    drb.setRouting(String.valueOf(routingKey));
-                }
-            }
-            drb.execute().actionGet();
-            entity.deleted = true;
-            queryDuration.addValue(w.elapsedMillis());
-            w.submitMicroTiming("ES", "DELETE " + entity.getClass().getName());
-            entity.afterDelete();
-            if (LOG.isFINE()) {
-                LOG.FINE("DELETE: %s.%s: %s SUCCESS",
-                         getIndex(entity.getClass()),
-                         descriptor.getType(),
-                         entity.getId());
-            }
-            traceChange(entity);
-        } catch (VersionConflictEngineException e) {
-            if (LOG.isFINE()) {
-                LOG.FINE("Version conflict on updating: %s", entity);
-            }
-            reportClash(entity);
-            optimisticLockErrors.inc();
-            throw new OptimisticLockException(e, entity);
-        } catch (Throwable e) {
-            throw Exceptions.handle()
-                            .to(LOG)
-                            .error(e)
-                            .withSystemErrorMessage("Failed to delete '%s' (%s): %s (%s)",
-                                                    entity.toDebugString(),
-                                                    entity.getId())
-                            .handle();
-        }
-    }
-
-    protected static <E extends Entity> void reportClash(E entity) {
-        if (!traceOptimisticLockErrors) {
-            return;
-        }
-        IndexTrace offendingTrace = traces.get(entity.getClass().getName() + "-" + entity.getId());
-        if (offendingTrace != null) {
-            StringBuilder msg = new StringBuilder("Detected optimistic locking error:\n");
-            msg.append("Current Thread: ");
-            msg.append(Thread.currentThread().getName());
-            msg.append("\n");
-            for (StackTraceElement e : Thread.currentThread().getStackTrace()) {
-                msg.append(e.getClassName())
-                   .append(".")
-                   .append(e.getMethodName())
-                   .append(" (")
-                   .append(e.getFileName())
-                   .append(":")
-                   .append(e.getLineNumber())
-                   .append(")\n");
-            }
-            msg.append("\n");
-            for (Tuple<String, String> t : CallContext.getCurrent().getMDC()) {
-                msg.append(t.getFirst()).append(": ").append(t.getSecond()).append("\n");
-            }
-            msg.append("\n");
-            msg.append("Offending Thread: ");
-            msg.append(offendingTrace.threadName);
-            msg.append("\n");
-            for (StackTraceElement e : offendingTrace.stackTrace) {
-                msg.append(e.getClassName())
-                   .append(".")
-                   .append(e.getMethodName())
-                   .append(" (")
-                   .append(e.getFileName())
-                   .append(":")
-                   .append(e.getLineNumber())
-                   .append(")\n");
-            }
-            msg.append("\n");
-            for (Tuple<String, String> t : offendingTrace.mdc) {
-                msg.append(t.getFirst()).append(": ").append(t.getSecond()).append("\n");
-            }
-            msg.append("\n");
-            LOG.SEVERE(msg.toString());
-        }
-    }
-
-    protected static <E extends Entity> void traceChange(E entity) {
-        if (!traceOptimisticLockErrors) {
-            return;
-        }
-        IndexTrace trace = new IndexTrace();
-        trace.threadName = Thread.currentThread().getName();
-        trace.id = entity.getId();
-        trace.type = entity.getClass().getName();
-        trace.timestamp = System.currentTimeMillis();
-        trace.mdc = CallContext.getCurrent().getMDC();
-        trace.stackTrace = Thread.currentThread().getStackTrace();
-        traces.put(trace.type + "-" + trace.id, trace);
+        Exceptions.logDeprecatedMethodUse();
+        access.forceDelete(entity);
     }
 
     /**
@@ -1398,118 +435,7 @@ public class Index {
      * @return a new query against the database
      */
     public static <E extends Entity> Query<E> select(Class<E> clazz) {
-        return new Query<>(clazz);
-    }
-
-    /**
-     * Installs the given ElasticSearch client
-     *
-     * @param client the client to use
-     */
-    public static void setClient(Client client) {
-        Index.client = client;
-    }
-
-    /**
-     * Returns the underlying ElasticSearch client
-     *
-     * @return the client used by this class
-     */
-    public static Client getClient() {
-        return client;
-    }
-
-    /**
-     * Returns the selected index prefix
-     *
-     * @return the index prefix added to each index name
-     */
-    public static String getIndexPrefix() {
-        return indexPrefix;
-    }
-
-    /**
-     * Sets the index prefix to use
-     *
-     * @param indexPrefix the index prefix to use
-     */
-    public static void setIndexPrefix(String indexPrefix) {
-        Index.indexPrefix = indexPrefix;
-    }
-
-    /**
-     * Generates a pure unclustered in memory instance of elasticsearch (most probably for testing purposes).
-     */
-    public static void generateEmptyInMemoryInstance() {
-        if (inMemoryNode != null) {
-            ready = false;
-            client.close();
-            inMemoryNode.close();
-        }
-
-        File tmpDir = new File(System.getProperty("java.io.tmpdir"), CallContext.getNodeName() + "_in_memory_es");
-        tmpDir.mkdirs();
-
-        Settings settings = ImmutableSettings.settingsBuilder()
-                                             .put("node.http.enabled", false)
-                                             .put("path.data", tmpDir.getAbsolutePath())
-                                             .put("index.gateway.type", "none")
-                                             .put("gateway.type", "none")
-                                             .put("index.store.type", "memory")
-                                             .put("index.number_of_shards", 1)
-                                             .put("index.number_of_replicas", 0)
-                                             .put("script.disable_dynamic", false)
-                                             .build();
-        inMemoryNode = NodeBuilder.nodeBuilder().data(true).settings(settings).local(true).node();
-        client = inMemoryNode.client();
-        if (schema != null) {
-            for (String msg : schema.createMappings()) {
-                LOG.FINE(msg);
-            }
-        }
-        ready = true;
-    }
-
-    @Part
-    private static Resources resources;
-
-    /**
-     * Loads a test dataset from the classpath. The given file should contains an array of JSON objects.
-     * <p>
-     * Each object must contain a property named <tt>_type</tt> which determines the target entity as well as
-     * <tt>_id</tt> which determiens its ID.
-     *
-     * @param dataset the resource to load into the in memory elasticsearch.
-     */
-    @SuppressWarnings("unchecked")
-    public static void loadDataset(String dataset) {
-        try {
-            if (inMemoryNode == null) {
-                throw Exceptions.createHandled()
-                                .withSystemErrorMessage("Cannot load datasets when not running as 'in-memory'")
-                                .handle();
-            }
-            LOG.INFO("Loading dataset: %s", dataset);
-            Resource res = resources.resolve(dataset)
-                                    .orElseThrow(() -> new IllegalArgumentException("Unknown dataset: " + dataset));
-            String contents = CharStreams.toString(new InputStreamReader(res.getUrl().openStream(), Charsets.UTF_8));
-            JSONArray json = JSON.parseArray(contents);
-            for (JSONObject obj : (List<JSONObject>) (Object) json) {
-                try {
-                    String type = obj.getString("_type");
-                    Class<? extends Entity> entityClass = Index.getType(type);
-                    EntityDescriptor descriptor = getDescriptor(entityClass);
-                    Entity entity = entityClass.newInstance();
-                    entity.setId(obj.getString("_id"));
-                    descriptor.readSource(entity, obj);
-                    update(entity);
-                } catch (Throwable e) {
-                    throw new IllegalArgumentException("Cannot load: " + obj, e);
-                }
-            }
-            blockThreadForUpdate();
-        } catch (IOException e) {
-            throw Exceptions.handle(e);
-        }
+        Exceptions.logDeprecatedMethodUse();
+        return access.select(clazz);
     }
 }
