+++ src/main/java/sirius/search/Query.java
@@ -42,6 +42,7 @@ import sirius.kernel.commons.Tuple;
 import sirius.kernel.commons.ValueHolder;
 import sirius.kernel.commons.Watch;
 import sirius.kernel.di.std.ConfigValue;
+import sirius.kernel.di.std.Part;
 import sirius.kernel.health.Exceptions;
 import sirius.kernel.health.Microtiming;
 import sirius.kernel.nls.NLS;
@@ -73,7 +74,7 @@ import java.util.function.Consumer;
 import java.util.function.Function;
 
 /**
- * Represents a query against the database which are created via {@link Index#select(Class)}.
+ * Represents a query against the database which are created via {@link IndexAccess#select(Class)}.
  *
  * @param <E> the type of entities being queried
  */
@@ -111,6 +112,9 @@ public class Query<E extends Entity> {
     // Used to signal that deliberately no routing was given
     private boolean deliberatelyUnrouted;
     private int scrollTTL = SCROLL_TTL_SECONDS;
+
+    @Part
+    private static IndexAccess indexAccess;
 
     /**
      * Used to create a nwe query for entities of the given class
@@ -304,7 +308,7 @@ public class Query<E extends Entity> {
                 result.add(Collections.singletonList(token));
             }
         } catch (IOException e) {
-            Exceptions.handle(Index.LOG, e);
+            Exceptions.handle(IndexAccess.LOG, e);
         }
 
         return result;
@@ -387,7 +391,7 @@ public class Query<E extends Entity> {
      * @return the query itself for fluent method calls
      */
     public Query<E> index(String indexToUse) {
-        this.index = Index.getIndexName(indexToUse);
+        this.index = indexAccess.getIndexName(indexToUse);
         return this;
     }
 
@@ -418,7 +422,7 @@ public class Query<E extends Entity> {
      * @return the query itself for fluent method calls
      */
     protected Query<?> autoRoute(String field, String value) {
-        EntityDescriptor descriptor = Index.getDescriptor(clazz);
+        EntityDescriptor descriptor = indexAccess.getDescriptor(clazz);
         if (!descriptor.hasRouting()) {
             return this;
         }
@@ -499,7 +503,7 @@ public class Query<E extends Entity> {
      * @return the query itself for fluent method calls
      */
     public Query<E> addTermFacet(String field, String value, ValueComputer<String, String> translator) {
-        final Property p = Index.getDescriptor(clazz).getProperty(field);
+        final Property p = indexAccess.getDescriptor(clazz).getProperty(field);
         if (p instanceof EnumProperty && translator == null) {
             translator = (v) -> String.valueOf(((EnumProperty) p).transformFromSource(v));
         }
@@ -579,7 +583,7 @@ public class Query<E extends Entity> {
      * @return the query itself for fluent method calls
      */
     public Query<E> addDateRangeFacet(String field, String value, DateRange... ranges) {
-        final Property p = Index.getDescriptor(clazz).getProperty(field);
+        final Property p = indexAccess.getDescriptor(clazz).getProperty(field);
 
         DateFacet dateFacet = new DateFacet(p.getFieldTitle(), field, value, ranges);
         termFacets.add(dateFacet);
@@ -703,15 +707,15 @@ public class Query<E extends Entity> {
                 return null;
             }
             SearchRequestBuilder srb = buildSearch();
-            if (Index.LOG.isFINE()) {
-                Index.LOG.FINE("SEARCH-FIRST: %s.%s: %s",
-                               Index.getIndex(clazz),
-                               Index.getDescriptor(clazz).getType(),
-                               buildQuery());
+            if (IndexAccess.LOG.isFINE()) {
+                IndexAccess.LOG.FINE("SEARCH-FIRST: %s.%s: %s",
+                                     indexAccess.getIndex(clazz),
+                                     indexAccess.getDescriptor(clazz).getType(),
+                                     buildQuery());
             }
             return transformFirst(srb);
         } catch (Throwable e) {
-            throw Exceptions.handle(Index.LOG, e);
+            throw Exceptions.handle(IndexAccess.LOG, e);
         }
     }
 
@@ -726,23 +730,25 @@ public class Query<E extends Entity> {
     }
 
     private SearchRequestBuilder buildSearch() {
-        EntityDescriptor ed = Index.getDescriptor(clazz);
-        SearchRequestBuilder srb = Index.getClient()
-                                        .prepareSearch(index != null ? index : Index.getIndexName(ed.getIndex()))
-                                        .setTypes(ed.getType());
+        EntityDescriptor ed = indexAccess.getDescriptor(clazz);
+        SearchRequestBuilder srb = indexAccess.getClient()
+                                              .prepareSearch(index != null ?
+                                                             index :
+                                                             indexAccess.getIndexName(ed.getIndex()))
+                                              .setTypes(ed.getType());
         srb.setVersion(true);
         if (primary) {
             srb.setPreference("_primary");
         }
 
-        applyRouting(ed, srb);
+        applyRouting(ed, srb::setRouting);
         applyOrderBys(srb);
         applyFacets(srb);
         applyQueriesAndFilters(srb);
         applyLimit(srb);
 
         if (logQuery) {
-            Index.LOG.INFO(srb);
+            IndexAccess.LOG.INFO(srb);
         }
 
         return srb;
@@ -803,28 +809,6 @@ public class Query<E extends Entity> {
         }
     }
 
-    private void applyRouting(EntityDescriptor ed, SearchRequestBuilder srb) {
-        if (Strings.isFilled(routing)) {
-            if (!ed.hasRouting()) {
-                Exceptions.handle()
-                          .to(Index.LOG)
-                          .withSystemErrorMessage("Performing a search on %s with a routing "
-                                                  + "- but entity has no routing attribute (in @Indexed)! "
-                                                  + "This will most probably FAIL!", clazz.getName())
-                          .handle();
-            }
-            srb.setRouting(routing);
-        } else if (ed.hasRouting() && !deliberatelyUnrouted) {
-            Exceptions.handle()
-                      .to(Index.LOG)
-                      .withSystemErrorMessage("Performing a search on %s without providing a routing. "
-                                              + "Consider providing a routing for better performance "
-                                              + "or call deliberatelyUnrouted() to signal that routing was "
-                                              + "intentionally skipped.", clazz.getName())
-                      .handle();
-        }
-    }
-
     /**
      * Executes the query and returns a list of all matching entities.
      * <p>
@@ -859,21 +843,21 @@ public class Query<E extends Entity> {
                 defaultLimitEnforced = true;
             }
             SearchRequestBuilder srb = buildSearch();
-            if (Index.LOG.isFINE()) {
-                Index.LOG.FINE("SEARCH: %s.%s: %s",
-                               Index.getIndex(clazz),
-                               Index.getDescriptor(clazz).getType(),
-                               buildQuery());
+            if (IndexAccess.LOG.isFINE()) {
+                IndexAccess.LOG.FINE("SEARCH: %s.%s: %s",
+                                     indexAccess.getIndex(clazz),
+                                     indexAccess.getDescriptor(clazz).getType(),
+                                     buildQuery());
             }
             ResultList<E> resultList = transform(srb);
             if (defaultLimitEnforced && resultList.size() == DEFAULT_LIMIT) {
-                Index.LOG.WARN("Default limit was hit when using Query.queryList or Query.queryResultList! "
-                               + "Please provide an explicit limit or use Query.iterate to remove this warning. "
-                               + "Query: %s, Location: %s", this, ExecutionPoint.snapshot());
+                IndexAccess.LOG.WARN("Default limit was hit when using Query.queryList or Query.queryResultList! "
+                                     + "Please provide an explicit limit or use Query.iterate to remove this warning. "
+                                     + "Query: %s, Location: %s", this, ExecutionPoint.snapshot());
             }
             return resultList;
         } catch (Throwable e) {
-            throw Exceptions.handle(Index.LOG, e);
+            throw Exceptions.handle(IndexAccess.LOG, e);
         }
     }
 
@@ -887,29 +871,11 @@ public class Query<E extends Entity> {
             if (forceFail) {
                 return 0;
             }
-            EntityDescriptor ed = Index.getDescriptor(clazz);
-            CountRequestBuilder crb = Index.getClient()
-                                           .prepareCount(index != null ? index : Index.getIndex(clazz))
-                                           .setTypes(ed.getType());
-            if (Strings.isFilled(routing)) {
-                if (!ed.hasRouting()) {
-                    Exceptions.handle()
-                              .to(Index.LOG)
-                              .withSystemErrorMessage("Performing a search on %s with a routing "
-                                                      + "- but entity has no routing attribute (in @Indexed)! "
-                                                      + "This will most probably FAIL!", clazz.getName())
-                              .handle();
-                }
-                crb.setRouting(routing);
-            } else if (ed.hasRouting() && !deliberatelyUnrouted) {
-                Exceptions.handle()
-                          .to(Index.LOG)
-                          .withSystemErrorMessage("Performing a search on %s without providing a routing. "
-                                                  + "Consider providing a routing for better performance "
-                                                  + "or call deliberatelyUnrouted() to signal that routing "
-                                                  + "was intentionally skipped.", clazz.getName())
-                          .handle();
-            }
+            EntityDescriptor ed = indexAccess.getDescriptor(clazz);
+            CountRequestBuilder crb = indexAccess.getClient()
+                                                 .prepareCount(index != null ? index : indexAccess.getIndex(clazz))
+                                                 .setTypes(ed.getType());
+            applyRouting(ed, crb::setRouting);
             QueryBuilder qb = buildQuery();
             if (qb != null) {
                 crb.setQuery(qb);
@@ -918,12 +884,38 @@ public class Query<E extends Entity> {
             if (sb != null) {
                 crb.setQuery(QueryBuilders.filteredQuery(qb, sb));
             }
-            if (Index.LOG.isFINE()) {
-                Index.LOG.FINE("COUNT: %s.%s: %s", Index.getIndex(clazz), ed.getType(), buildQuery());
+            if (IndexAccess.LOG.isFINE()) {
+                IndexAccess.LOG.FINE("COUNT: %s.%s: %s", indexAccess.getIndex(clazz), ed.getType(), buildQuery());
             }
             return transformCount(crb);
         } catch (Throwable t) {
-            throw Exceptions.handle(Index.LOG, t);
+            throw Exceptions.handle(IndexAccess.LOG, t);
+        }
+    }
+
+    private void applyRouting(EntityDescriptor ed, Consumer<String> routingTarget) {
+        if (Strings.isFilled(routing)) {
+            if (!ed.hasRouting()) {
+                Exceptions.handle()
+                          .to(IndexAccess.LOG)
+                          .withSystemErrorMessage("Performing a search on %s without providing a routing. "
+                                                  + "Consider providing a routing for better performance "
+                                                  + "or call deliberatelyUnrouted() to signal that routing "
+                                                  + "was intentionally skipped. Query: %s",
+                                                  clazz.getName(),
+                                                  this.toString())
+                          .handle();
+            }
+            routingTarget.accept(routing);
+        } else if (ed.hasRouting() && !deliberatelyUnrouted) {
+            Exceptions.handle()
+                      .to(IndexAccess.LOG)
+                      .withSystemErrorMessage("Performing a query on %s with a routing "
+                                              + "- but entity has no routing attribute (in @Indexed)! "
+                                              + "This will most probably FAIL! Query: %s",
+                                              clazz.getName(),
+                                              this.toString())
+                      .handle();
         }
     }
 
@@ -989,7 +981,7 @@ public class Query<E extends Entity> {
         Watch w = Watch.start();
         SearchResponse searchResponse = builder.execute().actionGet();
         ResultList<E> result = new ResultList<>(termFacets, searchResponse);
-        EntityDescriptor descriptor = Index.getDescriptor(clazz);
+        EntityDescriptor descriptor = indexAccess.getDescriptor(clazz);
         for (SearchHit hit : searchResponse.getHits()) {
             E entity = clazz.newInstance();
             entity.initSourceTracing();
@@ -998,12 +990,12 @@ public class Query<E extends Entity> {
             descriptor.readSource(entity, hit.getSource());
             result.getResults().add(entity);
         }
-        if (Index.LOG.isFINE()) {
-            Index.LOG.FINE("SEARCH: %s.%s: SUCCESS: %d - %d ms",
-                           Index.getIndex(clazz),
-                           Index.getDescriptor(clazz).getType(),
-                           searchResponse.getHits().totalHits(),
-                           searchResponse.getTookInMillis());
+        if (IndexAccess.LOG.isFINE()) {
+            IndexAccess.LOG.FINE("SEARCH: %s.%s: SUCCESS: %d - %d ms",
+                                 indexAccess.getIndex(clazz),
+                                 indexAccess.getDescriptor(clazz).getType(),
+                                 searchResponse.getHits().totalHits(),
+                                 searchResponse.getTookInMillis());
         }
         if (Microtiming.isEnabled()) {
             w.submitMicroTiming("ES", "LIST: " + toString(true));
@@ -1028,14 +1020,14 @@ public class Query<E extends Entity> {
             result.initSourceTracing();
             result.setId(hit.getId());
             result.setVersion(hit.getVersion());
-            Index.getDescriptor(clazz).readSource(result, hit.getSource());
+            indexAccess.getDescriptor(clazz).readSource(result, hit.getSource());
         }
-        if (Index.LOG.isFINE()) {
-            Index.LOG.FINE("SEARCH-FIRST: %s.%s: SUCCESS: %d - %d ms",
-                           Index.getIndex(clazz),
-                           Index.getDescriptor(clazz).getType(),
-                           searchResponse.getHits().totalHits(),
-                           searchResponse.getTookInMillis());
+        if (IndexAccess.LOG.isFINE()) {
+            IndexAccess.LOG.FINE("SEARCH-FIRST: %s.%s: SUCCESS: %d - %d ms",
+                                 indexAccess.getIndex(clazz),
+                                 indexAccess.getDescriptor(clazz).getType(),
+                                 searchResponse.getHits().totalHits(),
+                                 searchResponse.getTookInMillis());
         }
         if (Microtiming.isEnabled()) {
             w.submitMicroTiming("ES", "FIRST: " + toString(true));
@@ -1052,11 +1044,11 @@ public class Query<E extends Entity> {
     protected long transformCount(CountRequestBuilder builder) {
         Watch w = Watch.start();
         CountResponse res = builder.execute().actionGet();
-        if (Index.LOG.isFINE()) {
-            Index.LOG.FINE("COUNT: %s.%s: SUCCESS: %d",
-                           Index.getIndex(clazz),
-                           Index.getDescriptor(clazz).getType(),
-                           res.getCount());
+        if (IndexAccess.LOG.isFINE()) {
+            IndexAccess.LOG.FINE("COUNT: %s.%s: SUCCESS: %d",
+                                 indexAccess.getIndex(clazz),
+                                 indexAccess.getDescriptor(clazz).getType(),
+                                 res.getCount());
         }
         if (Microtiming.isEnabled()) {
             w.submitMicroTiming("ES", "COUNT: " + toString(true));
@@ -1129,7 +1121,7 @@ public class Query<E extends Entity> {
             if (forceFail) {
                 return;
             }
-            EntityDescriptor entityDescriptor = Index.getDescriptor(clazz);
+            EntityDescriptor entityDescriptor = indexAccess.getDescriptor(clazz);
             SearchResponse searchResponse = createScroll(entityDescriptor);
             try {
                 TaskContext ctx = TaskContext.get();
@@ -1164,7 +1156,7 @@ public class Query<E extends Entity> {
                                 }
                             }
                         } catch (Exception e) {
-                            Exceptions.handle().to(Index.LOG).error(e).handle();
+                            Exceptions.handle().to(IndexAccess.LOG).error(e).handle();
                         }
                     }
                     if (searchResponse.getHits().hits().length == 0) {
@@ -1175,15 +1167,19 @@ public class Query<E extends Entity> {
                 clearScroll(searchResponse);
             }
         } catch (Throwable t) {
-            throw Exceptions.handle(Index.LOG, t);
+            throw Exceptions.handle(IndexAccess.LOG, t);
         }
     }
 
     private void clearScroll(SearchResponse searchResponse) {
         try {
-            Index.getClient().prepareClearScroll().addScrollId(searchResponse.getScrollId()).execute().actionGet();
+            indexAccess.getClient()
+                       .prepareClearScroll()
+                       .addScrollId(searchResponse.getScrollId())
+                       .execute()
+                       .actionGet();
         } catch (Throwable e) {
-            Exceptions.handle(Index.LOG, e);
+            Exceptions.handle(IndexAccess.LOG, e);
         }
     }
 
@@ -1197,7 +1193,7 @@ public class Query<E extends Entity> {
                                   "A scroll query against elasticserach took too long to process its data! "
                                   + "The result is probably inconsistent! Query: %s",
                                   this)
-                          .to(Index.LOG)
+                          .to(IndexAccess.LOG)
                           .handle();
             }
         }
@@ -1205,18 +1201,18 @@ public class Query<E extends Entity> {
     }
 
     private SearchResponse scrollFurther(EntityDescriptor entityDescriptor, SearchResponse searchResponse) {
-        searchResponse = Index.getClient()
-                              .prepareSearchScroll(searchResponse.getScrollId())
-                              .setScroll(org.elasticsearch.common.unit.TimeValue.timeValueSeconds(scrollTTL))
-                              .execute()
-                              .actionGet();
-        if (Index.LOG.isFINE()) {
-            Index.LOG.FINE("SEARCH-SCROLL: %s.%s: SUCCESS: %d/%d - %d ms",
-                           Index.getIndex(clazz),
-                           entityDescriptor.getType(),
-                           searchResponse.getHits().hits().length,
-                           searchResponse.getHits().totalHits(),
-                           searchResponse.getTookInMillis());
+        searchResponse = indexAccess.getClient()
+                                    .prepareSearchScroll(searchResponse.getScrollId())
+                                    .setScroll(org.elasticsearch.common.unit.TimeValue.timeValueSeconds(scrollTTL))
+                                    .execute()
+                                    .actionGet();
+        if (IndexAccess.LOG.isFINE()) {
+            IndexAccess.LOG.FINE("SEARCH-SCROLL: %s.%s: SUCCESS: %d/%d - %d ms",
+                                 indexAccess.getIndex(clazz),
+                                 entityDescriptor.getType(),
+                                 searchResponse.getHits().hits().length,
+                                 searchResponse.getHits().totalHits(),
+                                 searchResponse.getTookInMillis());
         }
         return searchResponse;
     }
@@ -1224,9 +1220,9 @@ public class Query<E extends Entity> {
     private SearchResponse createScroll(EntityDescriptor entityDescriptor) {
         SearchRequestBuilder srb = buildSearch();
         if (!orderBys.isEmpty()) {
-            Index.LOG.WARN("An iterated query cannot be sorted! Use '.blockwise(...)'. Query: %s, Location: %s",
-                           this,
-                           ExecutionPoint.snapshot());
+            IndexAccess.LOG.WARN("An iterated query cannot be sorted! Use '.blockwise(...)'. Query: %s, Location: %s",
+                                 this,
+                                 ExecutionPoint.snapshot());
         }
         srb.setSearchType(SearchType.SCAN);
         srb.setFrom(0);
@@ -1234,8 +1230,11 @@ public class Query<E extends Entity> {
         // Otherwise we limit to 10 documents per shard...
         srb.setSize(routing != null ? MAX_SCROLL_RESULTS_FOR_SINGLE_SHARD : MAX_SCROLL_RESULTS_PER_SHARD);
         srb.setScroll(org.elasticsearch.common.unit.TimeValue.timeValueSeconds(scrollTTL));
-        if (Index.LOG.isFINE()) {
-            Index.LOG.FINE("ITERATE: %s.%s: %s", Index.getIndex(clazz), entityDescriptor.getType(), buildQuery());
+        if (IndexAccess.LOG.isFINE()) {
+            IndexAccess.LOG.FINE("ITERATE: %s.%s: %s",
+                                 indexAccess.getIndex(clazz),
+                                 entityDescriptor.getType(),
+                                 buildQuery());
         }
 
         return srb.execute().actionGet();
@@ -1289,11 +1288,11 @@ public class Query<E extends Entity> {
             limit = 512;
             while (true) {
                 SearchRequestBuilder srb = buildSearch();
-                if (Index.LOG.isFINE()) {
-                    Index.LOG.FINE("PAGED-SEARCH: %s.%s: %s",
-                                   Index.getIndex(clazz),
-                                   Index.getDescriptor(clazz).getType(),
-                                   buildQuery());
+                if (IndexAccess.LOG.isFINE()) {
+                    IndexAccess.LOG.FINE("PAGED-SEARCH: %s.%s: %s",
+                                         indexAccess.getIndex(clazz),
+                                         indexAccess.getDescriptor(clazz).getType(),
+                                         buildQuery());
                 }
                 ResultList<E> resultList = transform(srb);
                 for (E entity : resultList) {
@@ -1315,7 +1314,7 @@ public class Query<E extends Entity> {
                             }
                         }
                     } catch (Exception e) {
-                        Exceptions.handle().to(Index.LOG).error(e).handle();
+                        Exceptions.handle().to(IndexAccess.LOG).error(e).handle();
                     }
                 }
                 // Re-create entity the duplicator
@@ -1327,7 +1326,7 @@ public class Query<E extends Entity> {
                 }
             }
         } catch (Throwable e) {
-            throw Exceptions.handle(Index.LOG, e);
+            throw Exceptions.handle(IndexAccess.LOG, e);
         }
     }
 
@@ -1398,7 +1397,7 @@ public class Query<E extends Entity> {
                 return;
             }
             Watch w = Watch.start();
-            EntityDescriptor ed = Index.getDescriptor(clazz);
+            EntityDescriptor ed = indexAccess.getDescriptor(clazz);
             // If there are foreign keys we cannot use delete by query...
             if (ed.hasForeignKeys()) {
                 deleteByIteration();
@@ -1409,18 +1408,20 @@ public class Query<E extends Entity> {
                 w.submitMicroTiming("ES", "DELETE: " + toString(true));
             }
         } catch (Throwable e) {
-            throw Exceptions.handle(Index.LOG, e);
+            throw Exceptions.handle(IndexAccess.LOG, e);
         }
     }
 
     protected void deleteByQuery(EntityDescriptor ed) {
-        DeleteByQueryRequestBuilder builder = Index.getClient()
-                                                   .prepareDeleteByQuery(index != null ? index : Index.getIndex(clazz))
-                                                   .setTypes(ed.getType());
+        DeleteByQueryRequestBuilder builder = indexAccess.getClient()
+                                                         .prepareDeleteByQuery(index != null ?
+                                                                               index :
+                                                                               indexAccess.getIndex(clazz))
+                                                         .setTypes(ed.getType());
         if (Strings.isFilled(routing)) {
             if (!ed.hasRouting()) {
                 throw Exceptions.handle()
-                                .to(Index.LOG)
+                                .to(IndexAccess.LOG)
                                 .withSystemErrorMessage("Performing a delete on %s with a routing "
                                                         + "- but entity has no routing attribute (in @Indexed)! "
                                                         + "This will most probably FAIL!", clazz.getName())
@@ -1429,7 +1430,7 @@ public class Query<E extends Entity> {
             builder.setRouting(routing);
         } else if (ed.hasRouting() && !deliberatelyUnrouted) {
             throw Exceptions.handle()
-                            .to(Index.LOG)
+                            .to(IndexAccess.LOG)
                             .withSystemErrorMessage("Performing a delete on %s without providing a routing. "
                                                     + "Consider providing a routing for better performance "
                                                     + "or call deliberatelyUnrouted() to signal that routing was "
@@ -1440,8 +1441,8 @@ public class Query<E extends Entity> {
         if (qb != null) {
             builder.setQuery(qb);
         }
-        if (Index.LOG.isFINE()) {
-            Index.LOG.FINE("DELETE: %s.%s: %s", Index.getIndex(clazz), ed.getType(), buildQuery());
+        if (IndexAccess.LOG.isFINE()) {
+            IndexAccess.LOG.FINE("DELETE: %s.%s: %s", indexAccess.getIndex(clazz), ed.getType(), buildQuery());
         }
         builder.execute().actionGet();
     }
@@ -1450,7 +1451,7 @@ public class Query<E extends Entity> {
         ValueHolder<Throwable> error = ValueHolder.of(null);
         iterate(e -> {
             try {
-                Index.delete(e);
+                indexAccess.delete(e);
             } catch (Throwable ex) {
                 error.set(ex);
             }
