+++ src/main/java/sirius/search/properties/BooleanProperty.java
@@ -12,6 +12,7 @@ import sirius.kernel.di.std.Register;
 import sirius.kernel.health.Exceptions;
 import sirius.search.Entity;
 import sirius.search.Index;
+import sirius.search.IndexAccess;
 import sirius.web.http.WebContext;
 
 import java.lang.reflect.Field;
@@ -59,7 +60,7 @@ public class BooleanProperty extends Property {
                 field.set(entity, transformFromRequest(getName(), ctx));
             }
         } catch (IllegalAccessException e) {
-            Exceptions.handle(Index.LOG, e);
+            Exceptions.handle(IndexAccess.LOG, e);
         }
     }
