+++ src/main/java/sirius/search/properties/EnumListProperty.java
@@ -13,7 +13,7 @@ import sirius.kernel.commons.Value;
 import sirius.kernel.di.std.Register;
 import sirius.kernel.health.Exceptions;
 import sirius.search.Entity;
-import sirius.search.Index;
+import sirius.search.IndexAccess;
 import sirius.search.annotations.ListType;
 import sirius.web.http.WebContext;
 
@@ -91,7 +91,7 @@ public class EnumListProperty extends Property {
         try {
             field.set(entity, transformFromRequest(getName(), ctx));
         } catch (IllegalAccessException e) {
-            Exceptions.handle(Index.LOG, e);
+            Exceptions.handle(IndexAccess.LOG, e);
         }
     }
