+++ src/main/java/sirius/search/properties/ObjectListProperty.java
@@ -12,7 +12,7 @@ import sirius.kernel.di.std.Register;
 import sirius.kernel.health.Exceptions;
 import sirius.kernel.nls.NLS;
 import sirius.search.Entity;
-import sirius.search.Index;
+import sirius.search.IndexAccess;
 import sirius.search.annotations.ListType;
 import sirius.web.http.WebContext;
 
@@ -78,7 +78,7 @@ public class ObjectListProperty extends Property {
                             } catch (Throwable e) {
                                 Exceptions.handle()
                                           .error(e)
-                                          .to(Index.LOG)
+                                          .to(IndexAccess.LOG)
                                           .withSystemErrorMessage("Cannot load POJO field %s of %s: %s (%s)",
                                                                   innerField.getName(),
                                                                   toString())
@@ -90,7 +90,7 @@ public class ObjectListProperty extends Property {
                     } catch (Throwable e) {
                         Exceptions.handle()
                                   .error(e)
-                                  .to(Index.LOG)
+                                  .to(IndexAccess.LOG)
                                   .withSystemErrorMessage("Cannot load POJO in %s: %s (%s)", toString())
                                   .handle();
                     }
@@ -119,7 +119,7 @@ public class ObjectListProperty extends Property {
                         } catch (Throwable e) {
                             Exceptions.handle()
                                       .error(e)
-                                      .to(Index.LOG)
+                                      .to(IndexAccess.LOG)
                                       .withSystemErrorMessage("Cannot save POJO field %s of %s: %s (%s)",
                                                               innerField.getName(),
                                                               toString())
