+++ src/main/java/sirius/search/properties/StringMapProperty.java
@@ -12,6 +12,7 @@ import sirius.kernel.di.std.Register;
 import sirius.kernel.health.Exceptions;
 import sirius.search.Entity;
 import sirius.search.Index;
+import sirius.search.IndexAccess;
 import sirius.search.annotations.ListType;
 import sirius.web.http.WebContext;
 
@@ -66,7 +67,7 @@ public class StringMapProperty extends Property {
         try {
             field.set(entity, transformFromRequest(getName(), ctx));
         } catch (IllegalAccessException e) {
-            Exceptions.handle(Index.LOG, e);
+            Exceptions.handle(IndexAccess.LOG, e);
         }
     }
