+++ src/main/java/sirius/search/util/ReIndexCommand.java
@@ -8,8 +8,10 @@
 
 package sirius.search.util;
 
+import sirius.kernel.di.std.Part;
 import sirius.kernel.di.std.Register;
 import sirius.search.Index;
+import sirius.search.IndexAccess;
 import sirius.web.health.console.Command;
 
 /**
@@ -18,12 +20,16 @@ import sirius.web.health.console.Command;
  */
 @Register
 public class ReIndexCommand implements Command {
+
+    @Part
+    private IndexAccess index;
+
     @Override
     public void execute(Output output, String... params) throws Exception {
         if (params.length != 1) {
             output.line("Usage: reindex <newIndexPrefix>");
         } else {
-            Index.getSchema().reIndex(params[0]);
+            index.getSchema().reIndex(params[0]);
             output.line("Operation has started!");
         }
     }
