+++ src/main/java/sirius/search/util/UpdateMappingCommand.java
@@ -10,10 +10,12 @@ package sirius.search.util;
 
 import sirius.kernel.commons.Monoflop;
 import sirius.kernel.commons.Value;
+import sirius.kernel.di.std.Part;
 import sirius.kernel.di.std.Register;
 import sirius.search.Entity;
 import sirius.search.EntityDescriptor;
 import sirius.search.Index;
+import sirius.search.IndexAccess;
 import sirius.web.health.console.Command;
 
 /**
@@ -21,6 +23,10 @@ import sirius.web.health.console.Command;
  */
 @Register
 public class UpdateMappingCommand implements Command {
+
+    @Part
+    private static IndexAccess index;
+
     @Override
     public void execute(Output output, String... params) throws Exception {
         if (Value.indexOf(0, params).isEmptyString()) {
@@ -28,10 +34,9 @@ public class UpdateMappingCommand implements Command {
         } else {
             Class<? extends Entity> type = findTypeOrReportError(output, params[0]);
             if (type != null) {
-                EntityDescriptor ed = Index.getDescriptor(type);
+                EntityDescriptor ed = index.getDescriptor(type);
                 try {
-                    Index.addMapping(Index.getIndexPrefix() + ed.getIndex(),
-                                     type);
+                    index.addMapping(index.getIndexName(ed.getIndex()), type);
                     output.line("Mapping was updated...");
                 } catch (Exception e) {
                     output.line(e.getMessage());
@@ -42,11 +47,11 @@ public class UpdateMappingCommand implements Command {
     }
 
     protected static Class<? extends Entity> findTypeOrReportError(Output output, String typeName) {
-        Class<? extends Entity> type = Index.getType(typeName);
+        Class<? extends Entity> type = index.getType(typeName);
         if (type == null) {
             output.line("Unknown type: " + typeName);
             Monoflop mf = Monoflop.create();
-            for (String name : Index.getSchema().getTypeNames()) {
+            for (String name : index.getSchema().getTypeNames()) {
                 if (name.toLowerCase().contains(typeName.toLowerCase())) {
                     if (mf.firstCall()) {
                         output.line("Did you mean one of those: ");
